Docuquest/requirements.txt
langchain
flask
tiktoken
openai
pypdf
fiass-cpu
python-dotenv

